+++ app/src/main/kotlin/com/oss/megane/navigation/MeganeNavGraph.kt
@@ -1,0 +1,21 @@
+package com.oss.megane.navigation
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import com.oss.megane.ui.account.AccountScreen
+import com.oss.megane.ui.home.HomeScreen
+
+@Composable
+fun NavigationGraph(navController: NavHostController, modifier: Modifier = Modifier) {
+    NavHost(navController, startDestination = BottomNavItem.Home.route, modifier = modifier) {
+        composable(BottomNavItem.Home.route) {
+            HomeScreen()
+        }
+        composable(BottomNavItem.Account.route) {
+            AccountScreen()
+        }
+    }
+}
